requirements.txt
anyascii==0.1.7
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.8.2
black==20.8b1
certifi==2020.11.8
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.4
django-admin-autocomplete-filter==0.6.1
django-filter==2.4.0
django-location-field==2.1.0
django-map-widgets==0.3.0
django-modelcluster==5.1
django-taggit==1.3.0
django-tailwind==1.0.0
django-treebeard==4.4
djangorestframework==3.12.2
draftjs-exporter==2.1.7
drf-spectacular==0.14.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
geographiclib==1.50
geopy==2.1.0
html5lib==1.1
identify==1.5.13
idna==2.10
importlib-metadata==2.1.1
importlib-resources==3.3.0
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
jsonschema==3.2.0
l18n==2020.6.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nodeenv==1.5.0
openpyxl==3.0.6
packaging==20.8
pathspec==0.8.1
pep8-naming==0.11.1
Pillow==8.0.1
pluggy==0.13.1
pre-commit==2.10.0
protobuf==3.15.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
taggit-selectize==2.10.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.18
virtualenv==20.2.2
webencodings==0.5.1
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==3.4.0
django-activity-stream==0.10.0
django-colorfield==0.4.1
django-loginas==0.3.9
